requirements.txt
# Python dependencies
brotli==1.2.0
pillow==12.0.0
